[IMP] sale_stock: the safety margin is whole days
"Sales Safety Days" was a Float. A margin in days has no fractional part: its
only consumer is `timedelta(days=self.order_id.company_id.security_lead)` in
`_prepare_procurement_vals`, which will happily shift a procurement date by
0.7 of a day that no warehouse can act on. Having no `digits`, the field also
rendered with two decimals, so Inventory > Settings read "Move forward expected
delivery dates by 0.00 days".

Solution:
- Declare `res.company.security_lead` and the `res.config.settings` related
  field as Integer, with an integer default
- Reset it to `0` rather than `0.0` in `_onchange_use_security_lead`

Adapted from upstream 6ac707ad6d7e (master), which converted four day-based
fields across `sale`, `sale_stock`, `stock` and `purchase_stock`. Only
`security_lead` is taken: it is declared and consumed entirely inside this
module. `customer_lead` is deliberately left alone -- it is declared
`fields.Float` in `addons/sale/models/sale_order_line.py` and this module only
overrides it, so converting the override would leave two declarations of one
field disagreeing about its type. `days_to_purchase` and `horizon_days` are in
other modules. The product's own `sale_delay` is already Integer here
(`addons/stock/models/product_template.py`), which is what made this a
half-finished sweep in our tree rather than a new idea.

Upstream's view hunk (a `<br/>` in the settings label) is not taken; our
template is in `wizard/res_config_settings_views.xml` and does not have the
width problem that motivated it.

The existing column is converted in place by `-u`: `odoo/db/schema.py`
issues `ALTER COLUMN ... TYPE int4 USING security_lead::int4`. Confirmed on a
database installed before this change -- `res_company.security_lead` went from
`double precision` to `integer` with no warning and no `_moved0` fallback.
Note the two roundings differ: Postgres rounds on that cast, while the ORM
truncates (`Integer.convert_to_cache` is `int(value or 0)`), so a stored 2.7
migrates to 3 but a fresh write of 2.7 lands as 2. Neither matters for real
data here; every value our tests and defaults use is whole.

Proven by TestSecurityLeadIsWholeDays, red before and green after:
- test_the_margin_is_declared_as_whole_days on both models
- test_a_fractional_margin_does_not_survive_a_write: 2.7 reads back as 2
- test_the_column_really_became_an_integer reads `information_schema`, so a
  field declared Integer over a surviving float8 column cannot pass

Verification:
    /sale_stock:TestSecurityLeadIsWholeDays on a disposable database
    before: 4 failed, 0 error(s) of 3 tests   (one is a subtest pair)
    after:  0 failed, 0 error(s) of 3 tests

Upstream-Reference: 6ac707ad6d7e
Task ID: 31489

+++ addons/sale_stock/wizard/res_config_settings.py
@@ -4,7 +4,7 @@ from odoo import api, fields, models
 class ResConfigSettings(models.TransientModel):
     _inherit = "res.config.settings"
 
-    security_lead = fields.Float(
+    security_lead = fields.Integer(
         related="company_id.security_lead",
         string="Security Lead Time",
         readonly=False,
@@ -28,4 +28,4 @@ class ResConfigSettings(models.TransientModel):
     @api.onchange("use_security_lead")
     def _onchange_use_security_lead(self):
         if not self.use_security_lead:
-            self.security_lead = 0.0
+            self.security_lead = 0
